+++ traffic_portal/app/src/common/api/JobService.js
@@ -31,9 +31,8 @@ var JobService = function($http, ENV) {
 	};
 
 	this.createJob = function(job) {
-		return $http.post(ENV.api['root'] + 'user/current/jobs', job).then(
+		return $http.post(ENV.api['root'] + 'jobs', job).then(
 			function (result) {
-				console.info('new job created: ', result);
 				return result;
 			},
 			function (err) {