+++ traffic_portal/app/src/common/api/ParameterService.js
@@ -95,7 +95,7 @@ var ParameterService = function($http, locationUtils, messageModel, ENV) {
     };
 
     this.getCacheGroupUnassignedParams = function(cgId) {
-        return $http.get(ENV.api['root'] + 'cachegroups/' + cgId + '/unassigned_parameters').then(
+        return $http.get(ENV.api['legacy'] + 'cachegroups/' + cgId + '/unassigned_parameters').then(
             function (result) {
                 return result.data.response;
             },